requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==24.1.0
aiogram==2.15
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
anyio==4.4.0
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
babel==2.16.0
backports.zoneinfo==0.2.1
base==0.0.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
boto3==1.35.13
botocore==1.35.13
CacheControl==0.12.14
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==2.5.3
cryptography==43.0.1
cycler==0.12.1
deap==1.4.1
dnspython==2.3.0
docopt==0.6.2
exceptiongroup==1.2.2
Faker==28.4.1
fastapi==0.114.1
fbmessenger==6.0.0
filelock==3.15.4
fire==0.6.0
Flask==2.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httptools==0.6.1
huggingface-hub==0.24.7
humanfriendly==10.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.2.0
importlib_resources==6.4.0
itsdangerous==2.2.0
jax==0.4.13
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.4
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.5
libclang==18.1.1
locket==1.0.0
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mattermostwrapper==2.2
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.8
mtcnn==0.1.1
multidict==5.2.0
multitasking==0.0.11
narwhals==1.7.0
nltk==3.8.1
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.5.4.58
opt-einsum==3.3.0
pamqp==3.2.1
pandas==2.0.3
partd==1.4.1
peewee==3.17.6
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==4.3.2
pluggy==1.5.0
portalocker==2.10.1
prompt-toolkit==3.0.28
protobuf==4.23.3
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==1.10.9
pydeck==0.9.1
pydot==1.4.2
Pygments==2.18.0
PyJWT==2.9.0
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.1.1
pyreadline3==3.4.1
pyrsistent==0.20.0
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.2
questionary==1.10.0
randomname==0.1.5
redis==4.6.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
safetensors==0.4.5
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==0.22.1
scipy==1.10.1
sentry-sdk==1.14.0
simple-websocket==1.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack_sdk==3.31.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
sqlparse==0.4.4
starlette==0.38.5
streamlit==1.38.0
structlog==23.3.0
structlog-sentry==2.1.0
sympy==1.13.1
tabulate==0.9.0
tarsafe==0.0.4
tenacity==8.5.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.2.0
tifffile==2023.7.10
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.0
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.1
transformers==4.44.2
twilio==8.2.2
typing-utils==0.1.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.20
watchdog==4.0.1
wcwidth==0.2.13
web3==6.10.0
webencodings==0.5.1
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==2.0.3
wordcloud==1.9.2
wrapt==1.14.1
wsproto==1.2.0
xgboost==2.1.1
xlwt==1.3.0
yarl==1.9.2
yfinance==0.2.43
zipp==3.17.0
